rest_api/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
